union type TS
As an ex: you want to impletement a function "combine" which will add numbers if both inputs are numbers also will concatinate if one of the inputs or both inputs are string type. Here by using union type we can make it happen in TS. But to work around the error which will be thrown we can simply implement a runtime check with a if condition like I have used here. But it's not necessary to have one always.