"never" type
Ex: In this case I have thrown a custom error with a message and an error code. However when we throw an error it does not return anything at all. We can check it simply trying to console it. But like in the most cases it does not even print "undefined" even. So this is the great place for us to define it as "never" since the function never returns anything forever.